Move match unit tests
