+++ Slask.UnitTests/DomainTests/MatchTests/StartDateTimeTests/BracketStartDateTimeTests.cs
@@ -10,7 +10,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Xunit;
 
-namespace Slask.UnitTests.DomainTests.GroupTests.StartDateTimeTests
+namespace Slask.UnitTests.DomainTests.MatchTests.StartDateTimeTests
 {
     public class BracketStartDateTimeTests
     {
@@ -20,41 +20,12 @@ namespace Slask.UnitTests.DomainTests.GroupTests.StartDateTimeTests
         private readonly Tournament tournament;
         private readonly TournamentIssueReporter tournamentIssueReporter;
         private readonly BracketRound bracketRound;
-        private BracketGroup bracketGroup;
 
         public BracketStartDateTimeTests()
         {
             tournament = Tournament.Create("GSL 2019");
             tournamentIssueReporter = tournament.TournamentIssueReporter;
             bracketRound = tournament.AddBracketRound("Bracket round", 3) as BracketRound;
-        }
-
-        [Fact]
-        public void IssueIsReportedWhenStartDateTimeForMatchIsSetEarlierThanAnyMatchInPreviousRound()
-        {
-            List<string> playerNames = new List<string>() { "Maru", "Stork", "Taeja", "Rain", "Bomber", "FanTaSy", "Stephano", "Thorzain" };
-            bracketRound.SetPlayersPerGroupCount(4);
-            BracketRound secondBracketRound = tournament.AddBracketRound("Bracket round 2", 3, 2) as BracketRound;
-
-            foreach (string playerName in playerNames)
-            {
-                bracketRound.RegisterPlayerReference(playerName);
-            }
-
-            bracketGroup = bracketRound.Groups.First() as BracketGroup;
-            BracketNode finalNodeFromFirstRound = bracketGroup.BracketNodeSystem.FinalNode;
-
-            BracketGroup bracketGroupFromSecondRound = secondBracketRound.Groups.First() as BracketGroup;
-            BracketNode finalNodeFromSecondRound = bracketGroupFromSecondRound.BracketNodeSystem.FinalNode;
-
-            Match finalFromFirstRound = finalNodeFromFirstRound.Match;
-            Match finalFromSecondRound = finalNodeFromSecondRound.Match;
-            DateTime oneHourBeforeFinalFromFirstRound = finalFromFirstRound.StartDateTime.AddHours(-1);
-
-            finalFromSecondRound.SetStartDateTime(oneHourBeforeFinalFromFirstRound);
-
-            finalFromSecondRound.StartDateTime.Should().Be(oneHourBeforeFinalFromFirstRound);
-            tournamentIssueReporter.Issues.Should().HaveCount(1);
         }
 
         [Fact]
@@ -68,7 +39,7 @@ namespace Slask.UnitTests.DomainTests.GroupTests.StartDateTimeTests
                 bracketRound.RegisterPlayerReference(playerName);
             }
 
-            bracketGroup = bracketRound.Groups.First() as BracketGroup;
+            BracketGroup bracketGroup = bracketRound.Groups.First() as BracketGroup;
             BracketNode finalNode = bracketGroup.BracketNodeSystem.FinalNode;
 
             Match final = finalNode.Match;
@@ -98,7 +69,7 @@ namespace Slask.UnitTests.DomainTests.GroupTests.StartDateTimeTests
                 bracketRound.RegisterPlayerReference(playerName);
             }
 
-            bracketGroup = bracketRound.Groups.First() as BracketGroup;
+            BracketGroup bracketGroup = bracketRound.Groups.First() as BracketGroup;
             List<BracketNode> quarterfinalTier = bracketGroup.BracketNodeSystem.GetBracketNodesInTier(2);
 
             DateTime twoHoursEarlier = quarterfinalTier[0].Match.StartDateTime.AddHours(-2);
